Extract storage config to own file
